pyproject.toml
[tool.poetry]
name = "e_commerce_api"
version = "0.1.0"
description = ""
authors = ["yeomc"]

[tool.poetry.dependencies]
python = "^3.10"
tortoise-orm = {extras = ["asyncmy"], version = "^0.18.1"}
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.17.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-mail = "^1.0.2"
python-dotenv = "^0.19.2"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
